CV2-6018: expand CustomList in a new component to reuse it in article list
